manifest: Organize manifest parser in a single file
And create a header for manifest only functions and structs.
Functions from manifest.c that handles only manifest data should
be moved accordingly.
